Merged the Terrain Visualization by Martin Barbisch with our code.
This consited of deleting Ground Cover and Ground Floor since they were
no longer needed, adding the two png files from the which the terrain is
rendered. Changing the scale of the Model from 5 to 1. Adding all the
files with a T in front of their name. And adding much more code to
Visualizer.java

The program takes input from the commmand line to create a simple
visualization. Currently the rocket takes off from Hawaii and the
scaling is not quite right. It also has a slight problem with rendering
where it flashes. This is limited functionality and is meant mostly for
demonstrational purposes at this point.

Also I have not gone over all the code to remove uneeded code for our
situtation but I have done most of it. A complete overun of all the new
code should be one to test if there is any code that will not be needed
but that should be done after we have decided what functionality we
would like in the visualizer.